refactor: (ECA-96) moved the Add to Cart button into its own component
