+++ src/feature/product-overview/product-overview-component.tsx
@@ -1,13 +1,8 @@
 import { Euro, RussianRuble } from 'lucide-react';
-import { useState, useContext } from 'react';
 
-import { Button } from '@/components/ui/button';
 import { Spinner } from '@/components/ui/spiner';
-import AuthContext from '@/feature/auth/login/auth-provider';
-import { addToCart } from '@/feature/catalog/adding-to-cart/add-to-cart';
-import { cn } from '@/lib/utils';
 
-import { useCartStore } from '../catalog/adding-to-cart/use-cart-store';
+import { AddToCartButton } from '../catalog/add-to-cart-button/add-to-cart-button';
 import { DeleteFromCartButton } from '../catalog/delete-from-cart-button/delete-from-cart-button';
 
 import { ProductImages } from './product-images';
@@ -15,11 +10,6 @@ import { useProductOverview } from './use-product-overview';
 
 export default function ProductOverview({ productId }: { productId: string }) {
   const { isLoading, error, ...product } = useProductOverview(productId);
-  const [isAdding, setIsAdding] = useState(false);
-
-  const isInCart = useCartStore((state) => state.isInCart(productId));
-  const { IS_AUTHORIZED } = useContext(AuthContext);
-  const token = localStorage.getItem('ACCESS_TOKEN_KEY');
 
   if (isLoading) {
     return (
@@ -41,15 +31,6 @@ export default function ProductOverview({ productId }: { productId: string }) {
   }
 
   const PriceIcon = product.currencyCode === 'EUR' ? Euro : RussianRuble;
-
-  async function handleAddToCart() {
-    setIsAdding(true);
-    try {
-      await addToCart(productId, IS_AUTHORIZED, token);
-    } finally {
-      setIsAdding(false);
-    }
-  }
 
   return (
     <div className="mx-auto w-full max-w-7xl p-6">
@@ -84,20 +65,7 @@ export default function ProductOverview({ productId }: { productId: string }) {
           <p className="text-muted-foreground mb-6">{product.description}</p>
 
           <div className="mt-8 flex gap-4">
-            <Button
-              size="lg"
-              className={cn(
-                'flex-1 transition-colors',
-                isInCart ? 'bg-chart-3 text-card hover:bg-chart-3/80' : '',
-              )}
-              onClick={() => {
-                void handleAddToCart();
-              }}
-              disabled={isAdding || isInCart}
-            >
-              {isAdding ? 'Adding...' : isInCart ? 'In Cart' : 'Add to Cart'}
-            </Button>
-
+            <AddToCartButton productId={productId} />
             <DeleteFromCartButton productId={productId} variantId={product.masterVariant.id} />
           </div>
         </div>
